// src/app/models/local-pericia.ts
export class LocalPericia {
    codigo: any;
    numguia: any;
    tipo_local: any;
    nome_tipo_local: any;
    uf: any;
    cep: any;
    municipio: any;
    bairro: any;
    logradouro: any;
    complemento: any;
    ponto_referencia: any;
    descricao: any;
    latitude: any;
    longitude: any;
    quesitacoes: any;
    usuario: any;
    concluir: any;
    tipo_guia: string;
    constructor() {

    }
}